% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.templates-package.R
\docType{package}
\name{proj.templates}
\alias{proj.templates-package}
\alias{proj.templates}
\title{\code{\link{proj.templates}}}
\description{
Scientific project development and management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwardlavender/proj.templates/}
  \item Report bugs at \url{https://github.com/edwardlavender/proj.templatess/issues/}
}

}
\author{
\strong{Maintainer}: Edward Lavender \email{edward.lavender@eawag.ch} (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})

}
