% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-template.R
\name{use_template_gitignore}
\alias{use_template_gitignore}
\title{Use a template .gitignore file}
\usage{
use_template_gitignore(root = rprojroot::find_rstudio_root_file())
}
\arguments{
\item{root}{A character string that defines the project root.}
}
\value{
The function is called for the side effect of updating .gitignore. \code{invisible(NULL)} is returned.
}
\description{
This function updates a .gitignore file with standard additions.
}
\details{
This function runs \code{\link[usethis:git_vaccinate]{usethis::git_vaccinate()}} and adds the following to .gitignore:
\itemize{
\item data-raw/
\item data/*
\item !data/inst
\item doc/
\item fig/
\item *.html
\item *.pdf
\item *.docx
}

This function is designed to be called after \code{\link[=use_template_proj]{use_template_proj()}}.
}
\examples{
#### Set up example
# Create a temporary project
wd <- getwd()
proj <- file.path(tempdir(), "tmp")
usethis::create_project(path = proj, open = FALSE)
# Populate with standard 'proj.templates' files
# ... see `use_template_proj()` and associated functions
# Use git
setwd(proj)
usethis::use_git()

#### Example (1): Update .gitignore file
readLines(".gitignore")
use_template_gitignore(proj)
readLines(".gitignore")

setwd(wd)

}
\seealso{
\code{\link[=use_template_proj]{use_template_proj()}}, \code{\link[=use_template_gitignore]{use_template_gitignore()}}, \code{\link[=use_template_readme]{use_template_readme()}}, \code{\link[=use_template_script]{use_template_script()}}, \code{\link[=use_template_tree]{use_template_tree()}}
}
\author{
Edward Lavender
}
