% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-template.R
\name{use_template_proj}
\alias{use_template_proj}
\title{Use a template RStudio Project structure}
\usage{
use_template_proj(
  root = rprojroot::find_rstudio_root_file(),
  overwrite = FALSE
)
}
\arguments{
\item{root}{A character string that defines the project root.}

\item{overwrite}{A logical variable that defines whether or not to overwrite pre-existing R files (see Details). This does not apply to directories, which are never overwritten.}
}
\value{
The function is called for the side effect of creating files/folders. \code{invisible(NULL)} is returned.
}
\description{
This function creates standard set of folders/files within an established RStudio Project.
}
\details{
This function creates the following sequence of folders and files within the project:
\itemize{
\item data-raw/ for raw data
\item data/ for data
\itemize{
\item inst/ for \code{\link{proj.templates}} project-management files
}
\item R/ for \code{R} scripts (relating to data processing and analysis), including:
\itemize{
\item define_global_param.R, a template script for global parameter definitions
\item define_helpers.R, a template script for helper function definitions
}
\item dev/ for R scripts (relating to project development); namely:
\itemize{
\item 01-dev.R, which records project development
\item 02-clone.R, which facilitates project cloning
}
\item fig/ for figures
\item doc/ for documents
}
}
\examples{
#### Set up example
# Create a temporary project
proj <- file.path(tempdir(), "tmp")
usethis::create_project(path = proj, open = FALSE)

#### Example (1): Use a template project
list.files(proj)
use_template_proj(root = proj)
list.dirs(proj, full.names = FALSE)
list.files(proj, recursive = TRUE)

#### Clean up example
unlink(proj, recursive = TRUE)

}
\seealso{
\code{\link[=use_template_proj]{use_template_proj()}}, \code{\link[=use_template_gitignore]{use_template_gitignore()}}, \code{\link[=use_template_readme]{use_template_readme()}}, \code{\link[=use_template_script]{use_template_script()}}, \code{\link[=use_template_tree]{use_template_tree()}}
}
\author{
Edward Lavender
}
