% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-template.R
\name{use_template_readme}
\alias{use_template_readme}
\title{Use a template README file}
\usage{
use_template_readme(
  title = "README",
  author,
  email,
  root = rprojroot::find_rstudio_root_file(),
  open = rlang::is_interactive()
)
}
\arguments{
\item{title}{A character string that defines the README's title.}

\item{author}{A character string that defines the lead author (e.g., \code{"Edward Lavender"}).}

\item{email}{A character string that defines a contact email address.}

\item{root}{A character string that defines the project root.}

\item{open}{A logical variable that defines whether or not to open the template. This requires \code{rstudioapi} is available. If not, \code{open} is silently ignored.}
}
\value{
The function is called for the side effect of creating a README file. If a file is created, \code{invisible(TRUE)} is returned; otherwise, \code{invisible(FALSE)} is returned.
}
\description{
This function adds a standard template README file to a project.
}
\details{
This function creates a standard README as part of \code{\link{proj.templates}}'s \verb{use_template_*()} workflow. The README includes the following sections:
\itemize{
\item Introduction
\item Description
\itemize{
\item Dependencies
\item Directories
}
\item Instructions
\item Methods
\item Results
}

These sections assume that the project has been created following a standard analytical workflow, with the project using \code{renv} for dependency management, a standard set of directories and selected file-naming conventions. This can be tailored manually to suite project requirements.

Internally, the function calls \code{\link[usethis:use_readme_rmd]{usethis::use_readme_rmd()}}, which has several nice features (such as error checks and the addition of a git hook for .md and .Rmd files) and then updates the contents of the README with a standardised template structure, customised by user inputs.
}
\examples{
#### Set up example
# Create a temporary project
proj <- file.path(tempdir(), "tmp")
usethis::create_project(path = proj, open = FALSE)
# Populate with standard 'proj.templates' files
# ... see `use_template_proj()` and associated functions

#### Example (1): Use a template readme
use_template_readme(
  title = "proj.templates",
  author = "Edward Lavender",
  email = "el72@st-andrews.ac.uk",
  root = proj
)
list.files(proj)

}
\seealso{
\code{\link[=use_template_proj]{use_template_proj()}}, \code{\link[=use_template_gitignore]{use_template_gitignore()}}, \code{\link[=use_template_readme]{use_template_readme()}}, \code{\link[=use_template_script]{use_template_script()}}, \code{\link[=use_template_tree]{use_template_tree()}}
}
\author{
Edward Lavender
}
