% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-template.R
\name{use_template_script}
\alias{use_template_script}
\title{Use a template R script}
\usage{
use_template_script(file, open = rlang::is_interactive(), overwrite = FALSE)
}
\arguments{
\item{file}{A character string that defines the name of the file to be created.}

\item{open}{A logical variable that defines whether or not to open the template. This requires \code{rstudioapi} is available. If not, \code{open} is silently ignored.}

\item{overwrite}{A logical variable that defines whether or not to overwrite \code{file} if it already exists.}
}
\value{
The function is used for the side effect of generating a template R script.
}
\description{
This function generates a template R script.
}
\examples{
if (interactive()) use_template_script(file.path(tempdir(), "tmp.R"))
}
\seealso{
\code{\link[=use_template_proj]{use_template_proj()}}, \code{\link[=use_template_gitignore]{use_template_gitignore()}}, \code{\link[=use_template_readme]{use_template_readme()}}, \code{\link[=use_template_script]{use_template_script()}}, \code{\link[=use_template_tree]{use_template_tree()}}
}
\author{
Edward Lavender
}
